Refactor DownloadableFile to isolate Paperclip vs ActiveStorage behavior
